refactor: AddBot переименован в Platforms
