+++ src/components/platforms/create-bot/create-bot.module.scss
