+++ src/components/platforms/how-it-works/how-it-works.module.scss
